refactor(core): move core manager
