+++ core/src/CMakeLists.txt
@@ -4,8 +4,9 @@ set(CMAKE_AUTOMOC ON)
 set(LOGOS_CORE_SOURCES
     logos_core.cpp
     logos_core.h
-    core_manager.cpp
-    core_manager.h
+    core_manager/core_manager.cpp
+    core_manager/core_manager.h
+    core_manager/core_manager_interface.h
     ../interface.h
     ../plugin_registry.h
 )
