+++ core/src/core_manager/core_manager.h
