+++ core/src/core_manager/core_manager_interface.h
